requirements.txt
matplotlib
pyserial
scipy
